requirements.txt
-e bzr+ssh://bazaar.launchpad.net/~cf-charmers/charm-helpers/cloud-foundry/#egg=charmhelpers
Jinja2==2.7.2
MarkupSafe==0.19
PyYAML==3.11
Tempita==0.5.2
bzr==2.6.0
coverage==3.7.1
#launchpadlib==1.10.2
mock==1.0.1
netaddr==0.7.11
netifaces==0.8
nose==1.3.3
--allow-external python-apt
--allow-unverified python-apt
python-apt
python-mimeparse==0.1.4
#shelltoolbox==0.2.1
testtools==0.9.35
wheel==0.23.0
wsgiref==0.1.2
